+++ lib/features/inventory/presentation/widgets/items_tab.dart
@@ -474,6 +474,7 @@ class _ItemsTabState extends ConsumerState<ItemsTab> {
             child: SearchField(
               hintText: l10n.inventorySearchHint,
               onChanged: _onSearch,
+              onScan: _onSearch,
             ),
           ),
           Wrap(